Refactor DBase tests to enhance byte assertion and improve readability
- Added a new method `AssertBytes` in `DBaseTest` for comparing byte spans with enhanced error messaging.
- Introduced a `ReadWriteReturnResult` struct to encapsulate expected and actual byte spans in `DBaseTest<T>`.
- Updated `RoundtripHeader`, `RoundtripDescriptors`, and `RoundtripRecords` methods to utilize the new `AssertBytes` method for better clarity and maintainability.
- Changed the way byte ranges are asserted for DBase02 and other versions, ensuring proper handling of record counts and lengths.
- Modified the import in `DbfMarshalDateTimeTests` to reflect the new namespace structure for fields.
