+++ src/DBase/DbfFieldFlags.cs
@@ -1,29 +1,40 @@
 ﻿namespace DBase;
 
 /// <summary>
-/// Represents the flags of a <see cref="DbfField" />.
+/// Defines per-field attribute bits stored in a DBF field descriptor.
 /// </summary>
+/// <remarks>
+/// These values map to the flag byte in <see cref="DbfFieldDescriptor.Flags"/>. Flags are primarily used by
+/// Visual FoxPro variants and may be combined bitwise.
+/// </remarks>
 [Flags]
 public enum DbfFieldFlags : byte
 {
     /// <summary>
-    /// No flags.
+    /// No special attributes are set.
     /// </summary>
     None = 0x0,
+
     /// <summary>
-    /// System field.
+    /// Indicates a system-managed field.
     /// </summary>
     System = 0x1,
+
     /// <summary>
-    /// Nullable field.
+    /// Indicates the field can contain null values.
     /// </summary>
     Nullable = 0x2,
+
     /// <summary>
-    /// Binary field.
+    /// Indicates binary storage semantics for the field.
     /// </summary>
     Binary = 0x4,
+
     /// <summary>
-    /// Auto increment field.
+    /// Indicates an auto-increment field.
     /// </summary>
+    /// <remarks>
+    /// This is a composite value (<c>0x0C</c>) used by FoxPro-style descriptors.
+    /// </remarks>
     AutoIncrement = 0xC,
 }
