+++ src/DBase/DbfRecord.cs
@@ -5,8 +5,12 @@ using System.Diagnostics;
 namespace DBase;
 
 /// <summary>
-/// Represents a record of a dBASE file.
+/// Represents a single record (row) in a DBF table.
 /// </summary>
+/// <remarks>
+/// A DBF record starts with a one-byte delete flag followed by the field data, in the order defined by the
+/// table header. This type stores the persisted record status and materialized field values.
+/// </remarks>
 [DebuggerDisplay("{" + nameof(GetDebuggerDisplay) + "(),nq}")]
 public readonly record struct DbfRecord : IReadOnlyList<DbfField>
 {
@@ -14,12 +18,12 @@ public readonly record struct DbfRecord : IReadOnlyList<DbfField>
     internal ImmutableArray<DbfField> Fields { get; init; }
 
     /// <summary>
-    /// Gets a value indicating whether the record is deleted.
+    /// Gets a value indicating whether the record is marked as deleted in the DBF stream.
     /// </summary>
     public bool IsDeleted => Status is DbfRecordStatus.Deleted;
 
     /// <summary>
-    /// Gets the number of fields in the record.
+    /// Gets the number of fields in the record (as defined by the header).
     /// </summary>
     public int Count => Fields.Length;
 
@@ -28,13 +32,14 @@ public readonly record struct DbfRecord : IReadOnlyList<DbfField>
     /// </summary>
     /// <param name="index">The zero-based index of the field to get.</param>
     /// <returns>The <see cref="DbfField"/> at the specified index.</returns>
+    /// <exception cref="IndexOutOfRangeException"><paramref name="index"/> is outside record bounds.</exception>
     public DbfField this[int index] => Fields[index];
 
     /// <summary>
     /// Initializes a new instance of the <see cref="DbfRecord"/> record.
     /// </summary>
-    /// <param name="status">The record status.</param>
-    /// <param name="fields">The fields of the record.</param>
+    /// <param name="status">The persisted record status byte.</param>
+    /// <param name="fields">Field values in descriptor order.</param>
     public DbfRecord(DbfRecordStatus status, params ImmutableArray<DbfField> fields)
     {
         Status = status;
@@ -44,7 +49,7 @@ public readonly record struct DbfRecord : IReadOnlyList<DbfField>
     /// <summary>
     /// Initializes a new instance of the <see cref="DbfRecord"/> record.
     /// </summary>
-    /// <param name="fields">The fields of the record.</param>
+    /// <param name="fields">Field values in descriptor order.</param>
     public DbfRecord(params ImmutableArray<DbfField> fields)
     {
         Status = DbfRecordStatus.Valid;
@@ -56,7 +61,7 @@ public readonly record struct DbfRecord : IReadOnlyList<DbfField>
     /// <summary>
     /// Returns an enumerator that iterates through the fields.
     /// </summary>
-    /// <returns>An enumerator that can be used to iterate through the fields.</returns>
+    /// <returns>An enumerator that iterates fields in descriptor order.</returns>
     public IEnumerator<DbfField> GetEnumerator() => Fields.AsEnumerable().GetEnumerator();
 
     IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
@@ -76,9 +81,9 @@ public readonly record struct DbfRecord : IReadOnlyList<DbfField>
     }
 
     /// <summary>
-    /// Implicitly converts a DbfRecord instance to a read-only span of DbfField values, providing direct access to the
-    /// fields contained in the record.
+    /// Implicitly converts a <see cref="DbfRecord"/> to a read-only span of its fields.
     /// </summary>
-    /// <param name="record">The DbfRecord instance to convert. Must not be null.</param>
+    /// <param name="record">The record to convert.</param>
+    /// <returns>A span view over the field values.</returns>
     public static implicit operator ReadOnlySpan<DbfField>(DbfRecord record) => record.Fields.AsSpan();
 }
