+++ src/DBase/DbfVersion.cs
@@ -1,83 +1,83 @@
 ﻿namespace DBase;
 
 /// <summary>
-/// Represents the version of a DBF file.
+/// Identifies the DBF file-format version marker stored in the header.
 /// </summary>
 public enum DbfVersion : byte
 {
     /// <summary>
-    /// Unknown version.
+    /// Unspecified/unknown marker (<c>0x00</c>).
     /// </summary>
     Unknown = 0,
 
     /// <summary>
-    /// FoxBase version.
+    /// FoxBASE / dBASE II style table (<c>0x02</c>).
     /// </summary>
     DBase02 = 0x02,
 
     /// <summary>
-    /// FoxBase or dBASE III without memo fields.
+    /// FoxBASE or dBASE III table without memo fields (<c>0x03</c>).
     /// </summary>
     DBase03 = 0x03,
 
     /// <summary>
-    /// dBASE IV without memo fields.
+    /// dBASE IV table without memo fields (<c>0x04</c>).
     /// </summary>
     DBase04 = 0x04,
 
     /// <summary>
-    /// dBASE V without memo fields.
+    /// dBASE V table without memo fields (<c>0x05</c>).
     /// </summary>
     DBase05 = 0x05,
 
     /// <summary>
-    /// Visual FoxPro.
+    /// Visual FoxPro table (<c>0x30</c>).
     /// </summary>
     VisualFoxPro = 0x30,
 
     /// <summary>
-    /// Visual FoxPro with auto increment field.
+    /// Visual FoxPro table with auto-increment support (<c>0x31</c>).
     /// </summary>
     VisualFoxProWithAutoIncrement = 0x31,
 
     /// <summary>
-    /// Visual FoxPro with field type Varchar or Varbinary.
+    /// Visual FoxPro table with <c>Varchar</c>/<c>Varbinary</c> support (<c>0x32</c>).
     /// </summary>
     VisualFoxProWithVarchar = 0x32,
 
     /// <summary>
-    /// dBASE IV SQL table without memo fields.
+    /// dBASE IV SQL table without memo fields (<c>0x43</c>).
     /// </summary>
     DBase43 = 0x43,
 
     /// <summary>
-    /// dBASE IV SQL system file without memo fields.
+    /// dBASE IV SQL system file without memo fields (<c>0x63</c>).
     /// </summary>
     DBase63 = 0x63,
 
     /// <summary>
-    /// FoxBase or dBASE III with memo fields.
+    /// FoxBASE or dBASE III table with memo fields (<c>0x83</c>).
     /// </summary>
     DBase83 = 0x83,
 
     /// <summary>
-    /// dBASE IV with memo fields.
+    /// dBASE IV table with memo fields (<c>0x8B</c>).
     /// </summary>
     DBase8B = 0x8B,
 
     /// <summary>
-    /// dBASE IV SQL table with memo fields.
+    /// dBASE IV SQL table with memo fields (<c>0xCB</c>).
     /// </summary>
     // ReSharper disable once InconsistentNaming
     DBaseCB = 0xCB,
 
     /// <summary>
-    /// FoxPro 2.x with memo fields.
+    /// FoxPro 2.x table with memo fields (<c>0xF5</c>).
     /// </summary>
     FoxPro2WithMemo = 0xF5,
 
     /// <summary>
-    /// FoxBASE.
+    /// FoxBASE table marker (<c>0xFB</c>).
     /// </summary>
     // ReSharper disable once InconsistentNaming
     FoxBASE = 0xFB
