+++ src/DBase/Interop/Equality/DateEqualityComparer.cs
@@ -1,0 +1,15 @@
+﻿using System.Text;
+using DBase.Serialization.Fields;
+
+namespace DBase.Interop.Equality;
+
+internal sealed class DateEqualityComparer : DbfFieldEqualityComparer
+{
+    public static DateEqualityComparer Instance => field ??= new DateEqualityComparer();
+
+    public override bool Equals(byte[]? x, byte[]? y) =>
+        DbfFieldDateFormatter.ReadRaw(x, Encoding.UTF8) == DbfFieldDateFormatter.ReadRaw(y, Encoding.UTF8);
+
+    public override int GetHashCode(byte[]? obj) =>
+        DbfFieldDateFormatter.ReadRaw(obj, Encoding.UTF8).GetHashCode();
+}
