+++ src/DBase/Interop/Equality/DbfFieldEqualityComparer.cs
@@ -1,0 +1,32 @@
+﻿using System.Text;
+
+namespace DBase.Interop.Equality;
+
+internal abstract class DbfFieldEqualityComparer : IEqualityComparer<byte[]>
+{
+    public abstract bool Equals(byte[]? x, byte[]? y);
+    public abstract int GetHashCode(byte[]? obj);
+
+    public static IEqualityComparer<byte[]> Create(DbfFieldDescriptor descriptor, Encoding encoding) => descriptor.Type switch
+    {
+        DbfFieldType.AutoIncrement => BinaryEqualityComparer.Instance,
+        DbfFieldType.Binary => BinaryEqualityComparer.Instance,
+        DbfFieldType.Blob => BinaryEqualityComparer.Instance,
+        DbfFieldType.Character => new StringEqualityComparer(encoding),
+        DbfFieldType.Currency => BinaryEqualityComparer.Instance,
+        DbfFieldType.Date => DateEqualityComparer.Instance,
+        DbfFieldType.DateTime => BinaryEqualityComparer.Instance,
+        DbfFieldType.Double => BinaryEqualityComparer.Instance,
+        DbfFieldType.Float => NumericEqualityComparer.Instance,
+        DbfFieldType.Int32 => BinaryEqualityComparer.Instance,
+        DbfFieldType.Logical => LogicalEqualityComparer.Instance,
+        DbfFieldType.Memo => BinaryEqualityComparer.Instance,
+        DbfFieldType.NullFlags => BinaryEqualityComparer.Instance,
+        DbfFieldType.Numeric => NumericEqualityComparer.Instance,
+        DbfFieldType.Ole => BinaryEqualityComparer.Instance,
+        DbfFieldType.Picture => BinaryEqualityComparer.Instance,
+        DbfFieldType.Timestamp => BinaryEqualityComparer.Instance,
+        DbfFieldType.Variant => VariantEqualityComparer.Instance,
+        _ => throw new ArgumentOutOfRangeException(nameof(descriptor), descriptor, null)
+    };
+}
